added a profile to the start screen instead of a new page
