+++ veggie_fruit_growpros/src/App.js
@@ -4,7 +4,6 @@ import "./components/BrowsePage/BrowsePage.css";
 import './App.css';
 import Navbar from "./components/NavigationBar/Navbar";
 import Questionnaire from  "./components/Questionnaire/Questionnaire";
-import ProfilePage from './components/ProfilePage/ProfilePage';
 import {useState} from "react"
 
 
@@ -13,15 +12,15 @@ function App() {
   const [changePage, setChangePage] = useState("questionnaire");
   
 function changePageFunction(pagestring) {
-  setChangePage(pagestring)
+  setChangePage(pagestring)  
 }
 
   return (
     <div className="App">
       <Navbar changePageFunction={changePageFunction}/>
-      <BrowsePage changePage={changePage}/>
+      <BrowsePage changePage={changePage} setChangePage={setChangePage}/>
       <Questionnaire changePage={changePage} setChangePage={setChangePage}/>
-      <ProfilePage changePage={changePage}/>
+      
       {/* add a LandingPage here*/}
     </div>
   );
