+++ veggie_fruit_growpros/src/components/BrowsePage/ProfilePage/profile.css
@@ -1,0 +1,111 @@
+.ProfilePage{
+  display: flex;
+  flex-direction: column;
+  position: fixed;
+  height: 50vh;
+  width: 400px;
+  right: 0;
+  background-color: white;
+  transition: 0.7s ease-out;
+  
+}
+
+.profileCenter{
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content:space-evenly ;
+  width: 100%;
+  height: 100%;
+  background-color: white;
+  transition-delay: 1s;
+}
+
+.profile__topSection{
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+  width: 100%;
+  height: 30%;
+  
+}
+.profile__midSection{
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  align-items: center;
+
+  position: relative;
+  width: 100%;
+  height: 30%;
+  
+}
+.profile__BottomSection{
+  position: relative;
+  width: 100%;
+  height: 30%;
+  background-size: contain;
+  background-repeat: no-repeat;
+  background-position: center;
+  background-color: white;
+}
+.Profile__Circle__container{
+  height: 100px;
+  width: 100px;
+}
+
+.profile__Cirlce{
+  height: 90%;
+  width: 100%;
+  
+}
+.profile__Cirlce h5{
+  width: 100%;
+  height: 100%;
+  border-radius: 100px;
+  background-color: turquoise;
+  font-size: 4rem;
+}
+.profile__Bottom--title{
+  height: 20%;
+  width: 100%;
+}
+.cardImg{
+  height: 60%;
+  width: 100%;
+  background:no-repeat;
+  background-position: center;
+  background-size: contain;
+}
+
+.Exitbutton{
+  position: absolute;
+  right: 5px;
+  top: 10px;
+  z-index: 10;
+  background-color: white;
+  font-size: 2rem;
+  height: 44px;
+  width: 44px;
+  background-image: url(/src/components/images/xIcon.png);
+  background-position: center;
+  background-size: cover;
+}
+
+
+
+
+.NotProfile{
+  display: flex;
+  flex-direction: column;
+  position: absolute;
+  height: 0px;
+  width: 30%;
+  right: 50px;
+  
+  visibility: hidden;
+}
+
+
+
