--- veggie_fruit_growpros/src/components/ProfilePage/profile.css
@@ -1,45 +1,0 @@
-.ProfilePage{
-  height: 90vh;
-  max-width: 700px;
-  margin-left: auto;
-  margin-right: auto;
-  background-color: grey;
-}
-.ProfilePage__header{
-  height: 50%;
-  border-bottom: solid thin;
-}
-.ProfilePage__header__imgCont{
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 50%;
-  
-}
-.ProfilePage__header__imgCont__border{
-  
-  border: solid thick greenyellow;
-  width: 150px;
-  height: 150px;
-  border-radius: 100px;
-  text-align: center;
-  font-size: 7rem;
-}
-
-.ProfilePage__header__InfoContainer{
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  align-items: center;
-  height: 50%;
-}
-.ProfilePage__header__InfoContainer h3{
-  width: 50%;
-}
-
-
-
-.ProfilePage__main{
-  height: 50%;
-  width: 100%;
-}
